// api/src/__generated__/events.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: UBIf+m66m4BMZk+j9gLCF4dfsWPsPwY4X948//hmfjg2H8OEOmq4sA1W21rKyKlD5k4XPE1alk4s9ojpPRJn/g==
 */

// eslint:disable
// tslint:disable

interface Events {
  /**
   * @default CURRENT_TIMESTAMP
   */
  created_at: (Date) | null
  date: Date
  description: (string) | null
  facebook_event_link: (string) | null
  google_event_id: (string) | null
  /**
   * @default nextval('events_id_seq'::regclass)
   */
  id: number & {readonly __brand?: 'events_id'}
  location: (string) | null
  poster: (string) | null
  /**
   * @default true
   */
  show_on_frontpage: boolean
  title: string
  /**
   * @default CURRENT_TIMESTAMP
   */
  updated_at: (Date) | null
}
export default Events;

interface Events_InsertParameters {
  /**
   * @default CURRENT_TIMESTAMP
   */
  created_at?: (Date) | null
  date: Date
  description?: (string) | null
  facebook_event_link?: (string) | null
  google_event_id?: (string) | null
  /**
   * @default nextval('events_id_seq'::regclass)
   */
  id?: number & {readonly __brand?: 'events_id'}
  location?: (string) | null
  poster?: (string) | null
  /**
   * @default true
   */
  show_on_frontpage?: boolean
  title: string
  /**
   * @default CURRENT_TIMESTAMP
   */
  updated_at?: (Date) | null
}
export type {Events_InsertParameters}